requirements.txt
blinker==1.9.0
certifi==2025.11.12
click==8.3.1
Flask==3.1.3
fredapi==0.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
narwhals==2.18.0
numpy==2.4.3
nvidia-nccl-cu12==2.29.7
packaging==26.0
pandas==3.0.1
plotly==6.6.0
polygon-api-client==1.16.3
pyarrow==23.0.1
python-dateutil==2.9.0.post0
scipy==1.17.1
setuptools==82.0.1
six==1.17.0
urllib3==2.6.3
websockets==16.0
Werkzeug==3.1.6
wheel==0.46.3
xgboost==3.2.0
